requirements.txt
# Web Framework
streamlit>=1.28.0

# Data validation
pydantic>=2.5.0

# Data processing
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Optional (for better performance)
# ujson>=5.8.0
# orjson>=3.9.0
